requirements.txt
Package                      Version
---------------------------- --------
annotated-types              0.7.0
anyio                        4.4.0
asgiref                      3.7.2
autopep8                     2.3.1
cachetools                   5.3.3
certifi                      2024.6.2
charset-normalizer           3.3.2
click                        8.1.7
Django                       5.0
dnspython                    2.6.1
email_validator              2.2.0
fastapi                      0.111.0
fastapi-cli                  0.0.4
google-ai-generativelanguage 0.6.6
google-api-core              2.19.1
google-api-python-client     2.135.0
google-auth                  2.30.0
google-auth-httplib2         0.2.0
google-generativeai          0.7.1
googleapis-common-protos     1.63.2
grpcio                       1.64.1
grpcio-status                1.62.2
h11                          0.14.0
httpcore                     1.0.5
httplib2                     0.22.0
httptools                    0.6.1
httpx                        0.27.0
idna                         3.7
Jinja2                       3.1.4
markdown-it-py               3.0.0
MarkupSafe                   2.1.5
mdurl                        0.1.2
orjson                       3.10.5
pip                          24.0
proto-plus                   1.24.0
protobuf                     4.25.3
pyasn1                       0.6.0
pyasn1_modules               0.4.0
pycodestyle                  2.12.0
pydantic                     2.7.4
pydantic_core                2.18.4
Pygments                     2.18.0
pyparsing                    3.1.2
python-dotenv                1.0.1
python-multipart             0.0.9
PyYAML                       6.0.1
requests                     2.32.3
rich                         13.7.1
rsa                          4.9
setuptools                   65.5.0
shellingham                  1.5.4
sniffio                      1.3.1
sqlparse                     0.4.4
starlette                    0.37.2
tqdm                         4.66.4
typer                        0.12.3
typing_extensions            4.9.0
ujson                        5.10.0
uritemplate                  4.1.1
urllib3                      2.2.2
uvicorn                      0.30.1
uvloop                       0.19.0
watchfiles                   0.22.0
websockets                   12.0
